refactor (code base): refactored routing and pages for post related details
