+++ client/.eslintrc.json
@@ -35,6 +35,7 @@
     "linebreak-style": "off",
     "@typescript-eslint/no-use-before-define": ["error"],
     "react/jsx-wrap-multilines": 0,
+    "camelcase": 0,
     "react/jsx-filename-extension": [
       "warn",
       {
