+++ client/src/common/components/GlobalStyles.tsx
