+++ client/src/common/components/Main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import Navbar from '../common/components/Navbar';
-import GlobalStyles from '../styles/GlobalStyles';
+import Navbar from './Navbar';
+import GlobalStyles from './GlobalStyles';
 
 function Main() {
   return (
