+++ client/src/features/posts/components/PostDetails.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import '../index.css';
-import defaultIcon from '../assets/img/user.png';
+import defaultIcon from '../../../assets/img/default_user_image.png';
 
 const PostDetailsWrapper = styled.section`
   height: 70vh;
